+++ src/pages/HomeBestDetail/components/elements/BookMeta/BookMeta.style.ts
