+++ src/pages/HomeBestDetail/components/elements/EmotionBar/EmotionBar.style.ts
