+++ src/pages/HomeBestDetail/components/sections/BookReview/BookReivew.style.ts
@@ -1,7 +1,2 @@
 import { css } from '@emotion/react';
 import type { Theme } from '@emotion/react';
-
-export const cardWrapper = css`
-  display: flex;
-  gap: 1rem;
-`;
